can set the number of threads
git-svn-id: svn://svn.savannah.gnu.org/myserver/trunk@157 c4c922ea-a70d-0410-937d-ec1daf91936c

+++ myserverweb/source/cgi.cpp
@@ -258,7 +258,7 @@ int sendCGI(httpThreadContext* td,LPCONNECTION s,char* scriptpath,char* /*ext*/,
 		*/
 		sprintf(td->response.CONTENTS_DIM,"%u",nBytesRead-headerSize);
 		buildHTTPResponseHeader(td->buffer,&td->response);
-		ms_send(s->socket,td->buffer,strlen(td->buffer), 0);
+		ms_send(s->socket,td->buffer,(int)strlen(td->buffer), 0);
 		ms_send(s->socket,(char*)(td->buffer2+headerSize),nBytesRead-headerSize, 0);
 
 	}
